[FIX] dates and timezones : fix some bugs in tiki calendars item dates (e.g. in save, edit and show mode) and some date() that were not converted to gmdate().
More precisely, I removed the use of get_date_converter() that was using a bit buggy global variable. Instead, we now have two new functions/methods in tikilib :
	* date_UTC_to_display() : replacing getDisplayDateFromUTC()
	* date_display_to_UTC() : replacing getUTCFromDisplayDate()

("display" means "user local timezone stored in database" and "UTC" means "UTC+0 (i.e. GMT)")

Those functions are converting each dates with the user timezone name instead of the server offset, because the offset is not always the same with the same timezone name (it depends on the date too).

The timezone stuff in lib/smarty_tiki/modifier.tiki_date_format.php was also removed because we should not modify timezones (if possible) in smarty (it could be nice to always do this in tikilib) in order to avoid to convert date twice when using smarty modifiers over a date already passed through tikilib functions.

Tikilib date_format() function was modified to have no timezone modification (everything is now done in GMT instead of implicitely using the server timezone).

Since timestamps stored in the database are now in GMT timezone :
	* the calls to getDisplayDateFromServerDate() are replaced by date_UTC_to_display()
	* a new global variable "_DATE_TIMEZONE_DEFAULT" has been added in tiki-setup_base.php and set to "UTC" (this is used by the pear Date_TimeZone class)


git-svn-id: file:///var/local/code.tiki.org/tikisvn/trunk@7736 b456876b-0849-0410-b77d-98878d47e9d5