[FIX] calendar: fix a very annoying bug that was displaying a wrong ending hour for an event in editing mode.
More precisely, this ending hour was based on the 'time_end' that was not really taking care of timezones. the start hour was good since it was based on 'start' and not 'time_start'. The html_select_time fonction from smarty is able to manage timestamps (e.g. 'start' and 'end'), this is why it's not a problem to use them.

I'm completely removing the use of 'time_end' and 'time_start' from the source code, because they are not used anymore.
This fix also correctly handle ending and starting hours for a new event, that are separated by one hour.


git-svn-id: file:///var/local/code.tiki.org/tikisvn/trunk/tiki@8196 b456876b-0849-0410-b77d-98878d47e9d5