[NEW] search indexation: add a new indexation mode "normal" (the old one is named "random").
Note : This concerns only search indexation that doesn't use MySQL fulltext indexation.

* The "random" mode indexes tiki objects the following way :

	- each time a page (every php script that includes tiki-setup.php) is generated and sent to the browser, a random indexation is started,

	- this random indexation will do something only sometimes (more precisely, the function doesn't do anything if the result of a call to the php rand function returns 1 with a 1 to $search_refresh_rate range, defined in the search admin panel),

	- when it does it's job, it randomly choose one tiki object (i.e. one wiki page, file, forum comment, ...) and index it.

* The new "normal" generates it's indexes this way :

	- each time a tiki object content is created or modified, it is reindexed.

So, the "normal" mode is what is usually done everywhere, because it garanties that a new or modified content is immediately indexed. With the "random", you can, for example, upload files quicker, but you are never sure that all your content is indexed.

Since the "normal" mode seems to be what everyone usually expects, I choosed to set it up as the default mode (feel free to change this if you don't agree).


git-svn-id: file:///var/local/code.tiki.org/tikisvn/trunk/tiki@8261 b456876b-0849-0410-b77d-98878d47e9d5