[ENH] Use File Galleries for images inclusion in FCKEditor and normal page editor
Add a feature to allow the image selection by naviguating in the File Galleries,
both for the normal non WYSIWYG editor and FCKEditor.

You select the new feature in the admin page :
  "Use File Galleries for images inclusion"

Warning : The permission on the image(s) must be correct or users wont be able
to see the image(s).

There is a problem when generating PDF file from the page as
tiki-export_pdf.php is not using the users credentials when accessing the
elements of the page :(


git-svn-id: file:///var/local/code.tiki.org/tikisvn/trunk/tiki@9307 b456876b-0849-0410-b77d-98878d47e9d5