[MOD] tiki-setup: reorganize and divide tiki-setup in multiple smaller pieces
Tiki-setup.php is now requiring code only when needed in files placed in lib/setup (as done before with lib/setup/prefs.php). This is mainly done to have an easier to understand what is done in tiki-setup and to group parts of code that are related (instead of mixing everything). This should also reduce a bit the overhead, since files in lib/setup are mostly required if the features using them are activated.

This commit also includes some tiny changes :
  * change byref_set_value() to use simple_set_value(), because both use refs now,
  * auto-detect prefs that needs to be serialized/unserialized based on default values (should avoid missing prefs and should be more transparent for developpers), because it's done at another level (in tikilib and in lib/setup/prefs) => so, don't use serialized values in lib/setup/prefs.php
  * small typo in headerlib (matatags -> metatags)
  * remove some optional php closing tags to reduce risks (e.g. headers sent too early due to carriage return after the closing tag)
  * remove some unuseful smarty assignations and calls to get_preference()
  * don't use $prefs. prefix for 'title' (for the moment it's also assigned as $title)
  * assign headerlib in lib/headerlib.php, as it's done for others libs


git-svn-id: file:///var/local/code.tiki.org/tikisvn/trunk@9595 b456876b-0849-0410-b77d-98878d47e9d5