[MOD] File Galleries: Convert the root gallery (was gallery id '-1') of File Galleries from a fake gallery to a 'real' gallery in the database, in order to allow users to:
  - upload files directly into the root/main gallery, without having to create a subgallery first (this has mainly been done to reduce the steps needed to quickly upload an image in a wiki page - it should now be easier than the old wiki_up way)

  - change display properties of the root/main gallery

The ID of the main gallery is stored in the pref 'fgal_root_id', because it will be '1' only on new installs.
For the moment, the 'fgal_root_id' pref is used when calling functions from the filegallib and not hardcoded/forced inside those functions in order to be able to reuse those functions later for something else. To be more precise, the idea is to make it possible to store, for example, User Files inside the filegals table (in order to factorize code and merge both features).

+ Also factorize code of modules/mod-file_galleries.php and templates/tiki-list_file_gallery.tpl. Move a lot of code to filegallib in order to do so.



git-svn-id: file:///var/local/code.tiki.org/tikisvn/trunk@20826 b456876b-0849-0410-b77d-98878d47e9d5