[MOD] feature_wiki_paragraph_formatting_add_br has been removed to improve support for wysiwyg. This option is required when wysiwyg-wiki is running, but cannot be set when wysiwyg-html is active. This is confusing for the user. So, it seems best to remove it and let the system set it automatically.
There is a case when the pref value must be forced. (Done in wikiplugin_colorbox).
This is supported by a new option: 'formatting_add_br' in the call to tikilib parse_data.
The value can be set to one of: '' or 'y' or 'n'.
'' means it will use the default, which is 'y' when wysiwyg-wiki is active, and 'n' if wysiwyg-html is active


git-svn-id: file:///var/local/code.tiki.org/tikisvn/trunk@46090 b456876b-0849-0410-b77d-98878d47e9d5